requirements.txt
# Core dependencies - pinned to working versions
agno==1.7.7
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-identity>=1.15.0
certifi==2025.7.14
click==8.2.2
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.45
graphiti-core>=0.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mcp==1.12.3
mdurl==0.1.2
openai==1.98.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
rich==14.1.0
rpds-py==0.26.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0

# Ollama support
ollama>=0.1.8

# Optional dependencies for development/testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0

# Neo4j dependencies (may be needed for graphiti-core)
# neo4j>=5.0.0

# Additional runtime dependencies that might be needed
# fastapi>=0.100.0  # If using FastAPI features
# uvloop>=0.17.0    # For better async performance on Unix systems
